Add tests and fix bugs found while testing
