+++ tox.ini
@@ -5,6 +5,12 @@ envlist = py26,py27,py32,py33,py34,pypy,{py27,py34}-flake8
 pip_pre = False
 deps =
     -rdev-requirements.txt
+commands =
+    py.test {posargs:--cov rfc3986 tests/}
+    coverage report --fail-under 100 -m
+
+[testenv:pypy]
+deps = {[testenv]deps}
 commands = py.test {posargs}
 
 [testenv:py27-flake8]
